Improve Windows binutils activation for Autotools builds
Standard autotools builds check for a bunch of additional binutils
programs that we provide under special names, namely: as, dlltool, nm,
objdump, and strip. Provide additional environment variables for these.
This fixes libtool's ability to figure out which static libraries are
DLL import stubs on Windows, as work around in
https://github.com/conda-forge/xorg-libxdmcp-feedstock/pull/14 .