Remove `-ftree-vectorize` flag
Remove the `-ftree-vectorize` flag from the default flags.  It should
be unnecessary for optimized builds, as `-O2` already imples it
in the recent compiler versions, and it is undesirable for debug builds,
where it may introduce changes that make following the debugged code
harder.

For optimized builds, this "changes `-fvect-cost-model` from the `-O2`
default (`very-cheap`) to `cheap`", but that does not seem to have been
the original intention and it is unlikely to justify passing it
explicitly.

Fixes #165

Signed-off-by: Michał Górny <mgorny@quansight.com>